strip synthetic events from action creator args in dev mode
Our pattern of passing the underlying event as the last argument of
event handlers was causing generated action creators to get called with
the event, which at some point gets traversed, and the property access
throws a warning synthetic event pooling warning.

The events are never actually used by any of the reducers, so the
warning is benign afaict, so I'm only stripping them in dev mode.

The code is a little uglier than it could be because we're not
completely consistent about where the event goes (sometimes the last arg
is an object `{ props, event }`, sometimes the event itself).